Update speech recognition. 5k words, should work decently
