--- Mycroft-Core/src/main/assets/sync/en-us-ptm/README.md5
@@ -1,1 +1,0 @@
-45c80a79247b52dc6b48f2f07744d91d
