+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/feat.params
@@ -6,7 +6,7 @@
 -feat 1s_c_d_dd
 -svspec 0-12/13-25/26-38
 -agc none
--cmn batch
+-cmn current
 -varnorm no
 -model ptm
--cmninit 41.00,-5.29,-0.12,5.09,2.48,-4.07,-1.37,-1.78,-5.08,-2.05,-6.45,-1.42,1.17
+-cmninit 40,3,-1
