+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/feat.params.md5
@@ -1,1 +1,1 @@
-2dd4b7429d418f8fd42f5905570b77a6
+b2c57862cdae48b3a3970432ed26f31a
