+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/mdef.md5
@@ -1,1 +1,1 @@
-ba13d30c2fee63e039e119ab449bd618
+d31540bd4506dea2e89af493e649a616
