+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/noisedict
