+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/sendump.md5
