+++ Mycroft-Core/src/main/assets/sync/en-us-ptm/transition_matrices
