--- Mycroft-Core/src/main/assets/sync/en-us.lm.bin.md5
@@ -1,1 +1,0 @@
-574e3e017b3451d571a81c86f26128b7
