+++ Mycroft-Core/src/main/assets/sync/lm_giga_5k_nvp_2gram.lm.bin.md5
@@ -1,0 +1,1 @@
+473a30e383ee5f9ea5b3a62a979cc035  lm_giga_5k_nvp_2gram.lm.bin
