+++ Mycroft-Core/src/main/java/www/mabase/tech/mycroft/mycroftParser/TextDataCleaner.java
@@ -1,0 +1,24 @@
+package www.mabase.tech.mycroft.mycroftParser;
+
+public class TextDataCleaner {
+    /**
+     * Take all of the sample sentences that are saved in the intent database
+     */
+
+    void makeDictionary() {
+    //If there is not an existing dictionary
+    }
+
+    void updateDictionary(){
+        //If a new intent is installed, and it has unknown words
+        //or if the utterance has unknown words
+    }
+
+    void removeCommonWords(){
+        //This functionality may already be provided by DL4J, so standby
+    }
+
+    public void prepareUtterance(String utterance){
+        //Cleanup and prepare the utterance string
+    };
+}
