--- Mycroft-Core/src/main/java/www/mabase/tech/mycroft/mycroftparser/GenericClassifier.java
@@ -1,10 +1,0 @@
-package www.mabase.tech.mycroft.mycroftparser;
-
-public class GenericClassifier {
-    /**
-     * This exists as a special use case API. It is supposed to take serialized information over
-     * Androids IPC (Binder) that will corrispond to a NN model and dataset. This means an app
-     * can use ML without needing to package the library with their app. However, there should be
-     * a way to test this without forcing the developer to use Android studio, as that can be
-     */
-}
